Address Mark comments: 4

+++ src/Network/Network/Models/PSConnectionMonitorHttpConfiguration.cs
@@ -11,7 +11,7 @@ namespace Microsoft.Azure.Commands.Network.Models
     public class PSConnectionMonitorHttpConfiguration : PSNetworkWatcherConnectionMonitorProtocolConfiguration
     {
         [Ps1Xml(Target = ViewControl.Table)]
-        public int? Port { get; set; }
+        public short? Port { get; set; }
 
         [Ps1Xml(Target = ViewControl.Table)]
         public string Method { get; set; }
